refactor: #142 CreateFolderButtonをSharesに移動
