+++ src/components/post/Item/FolderList.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import { AiFillFolder as AiFillFolderIcon } from 'react-icons/ai'
 
-import { CreateFolderButton } from 'components/bookmark/CreateFolderButton'
+import { CreateFolderButton } from 'components/shares/button'
 import { useGetApi } from 'hooks/useApi'
 import { Folder } from 'types/bookmark'
 import { Post } from 'types/post'
