+++ src/components/shares/button/index.ts
@@ -1,2 +1,3 @@
 export { Button } from './Button'
 export { JumpToCreatePostButton } from './JumpToCreatePostButton'
+export { CreateFolderButton } from './CreateFolderButton'
