+++ src/pages/bookmark.tsx
@@ -2,9 +2,9 @@ import { NextPage } from 'next'
 import { useSearchParams } from 'next/navigation'
 import { useMemo } from 'react'
 import { BookmarkFolderList } from 'components/bookmark/BookmarkFolderList'
-import { CreateFolderButton } from 'components/bookmark/CreateFolderButton'
 import { Layout } from 'components/layout/Layout'
 import { PostItem } from 'components/post/Item/PostItem'
+import { CreateFolderButton } from 'components/shares/button'
 import { useRequireLogin } from 'hooks/login/useRequireLogin'
 import { useGetApi } from 'hooks/useApi'
 import { Folder, BookmarkPosts } from 'types/bookmark'
